fix call to sdscatfmt

+++ hiredis.c
@@ -608,7 +608,7 @@ long long redisFormatSdsCommandArgv(sds *target, int argc, const char **argv,
     cmd = sdscatfmt(cmd, "*%i\r\n", argc);
     for (j=0; j < argc; j++) {
         len = argvlen ? argvlen[j] : strlen(argv[j]);
-        cmd = sdscatfmt(cmd, "$%U\r\n", len);
+        cmd = sdscatfmt(cmd, "$%U\r\n", (unsigned long long)len);
         cmd = sdscatlen(cmd, argv[j], len);
         cmd = sdscatlen(cmd, "\r\n", sizeof("\r\n")-1);
     }
